+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+- optim: MARS
+- optim: common file for repeated ops (e.g., newton_schulz)
+
 ## [v0.3.0](https://github.com/stockeh/mlx-optimizers/releases/tag/v0.3.0) - 2024-11-18
 
 ### Added